requirements.txt
selenium>=4.10.0
python-dotenv>=1.0.0

selenium
webdriver_manager
beautifulsoup4
lxml
python-dotenv
